src/tagtry/dblayout/Generator.java
package tagtry.dblayout;

import java.util.Random;

import tagtry.integration.test.gender;

public class Generator {

	public static String randomstring() {

		char[] chars = "abcdefghijklmnopqrstuvwxyz".toCharArray();
		StringBuilder sb = new StringBuilder();
		Random random = new Random();
		for (int i = 0; i < 20; i++) {
			char c = chars[random.nextInt(chars.length)];
			sb.append(c);
		}
		return sb.toString();
	}

	public static int randomint() {
		return new Random().nextInt();
	}

	public static String randomemail() {
		char[] chars = "abcdefghijklmnopqrstuvwxyz".toCharArray();
		StringBuilder sb = new StringBuilder();
		Random random = new Random();
		for (int i = 0; i < 10; i++) {
			char c = chars[random.nextInt(chars.length)];
			sb.append(c);
		}
		sb.append("@");
		for (int i = 0; i < 3; i++) {
			char c = chars[random.nextInt(chars.length)];
			sb.append(c);
		}
		sb.append(".");

		for (int i = 0; i < 3; i++) {
			char c = chars[random.nextInt(chars.length)];
			sb.append(c);
		}

		return sb.toString();
	}

	public static String randomnumber() {

		char[] chars = "0123456789".toCharArray();
		StringBuilder sb = new StringBuilder();
		Random random = new Random();
		for (int i = 0; i < 10; i++) {
			char c = chars[random.nextInt(chars.length)];
			sb.append(c);
		}
		return sb.toString();
	}

	public static gender randomgender() {
		gender g = null;
		Random rand = new Random();
		boolean value = rand.nextBoolean();
		System.out.println(value);

		if (value == true) {
			g = gender.MAN;
		} else if (value == false) {
			g = gender.WOMEN;
		}
		return g;
	}
}
